requirements.txt
langchain
langchain-huggingface
sentence-transformers
chromadb
pymupdf
python-dotenv
